Fix #43 - Search page UI responsiveness sluggish when many spans in results (#46)
* update eslint-plugin-react, avoid eslintrc warnings

* .eslintrc env for browser, jest and jasmine

* Fix #43 - Search page responsiveness issues

Fix an issue with the search page that shows up when there are many
spans (cumulatively) in the search results. Issue at hand is the
mapStateToProps function calls .toJS() on some of the immutableJS state
values, transforms the resulting values, then passes the transformed
values to reselect. The .toJS() invocation prevents the calculations
from being memoized by reselect because it always results in new objects
or arrays.

The change is to wrap the .toJS() call and subsequent transformation in
a function, the result of which is cached based on the state value. The
caching resolves the issue. Only the last result needs to be cached
because these are top-level state values (vs transforming items in an
array) and therefore a state value will never be revisited.

* Fix yarn lock file, e.g. use npm registry

* Fix typo in comment

Signed-off-by: vvvprabhakar <vvvprabhakar@gmail.com>