Add more units to duration formatting
Expands the number of units that the formatDuration function displays.
Previously the function only used milliseconds and seconds to display
durations, now it will support all units up to days. This is to make
it easier to read very long durations in the UI that would have
previously only been displayed in seconds.

As a side effect of this change, there are a few changes to the way existing
durations are displayed:
* 0ms -> 0μs
* 0.15ms -> 150μs
* 1.7s -> 1s 700ms

Resolves #319

Signed-off-by: James Ferguson <jamesferguson497@gmail.com>