Improve human-readable duration formatting (#647)
* Add more units to duration formatting

Expands the number of units that the formatDuration function displays.
Previously the function only used milliseconds and seconds to display
durations, now it will support all units up to days. This is to make
it easier to read very long durations in the UI that would have
previously only been displayed in seconds.

As a side effect of this change, there are a few changes to the way existing
durations are displayed:
* 0ms -> 0μs
* 1.5μs
* 1.5ms
* 1.5s
* 1m 30s
* 1h 30m
* 1d 12h

Resolves #319

Signed-off-by: James Ferguson <jamesferguson497@gmail.com>

* Remove inputUnit parameter of formatDuration

The inputUnit parameter was not used anywhere and significantly
complicated the implementation of the formatDuration function.
Now, all inputs are assumed to be in microseconds which is consistent
with a few other methods in the date.tsx file.

Signed-off-by: James Ferguson <jamesferguson497@gmail.com>

* Reduce date format units to 2

Reduces the number of units displayed by the dateFormat function to
a maximum of 2. So "2d 5h 20m" will now be "2d 5h".

Signed-off-by: James Ferguson <jamesferguson497@gmail.com>

* Use decimals when formating μs, ms and s durations

This change will display durations less than a minute using decimals
as 1.3s is easier to comprehend than 1s 300ms as it was being
displayed before. Durations larger than a minute will still use multiple
units. Eg. 5m 27s

Signed-off-by: James Ferguson <jamesferguson497@gmail.com>

* Fix date.test.js licence

Signed-off-by: James Ferguson <jamesferguson497@gmail.com>