Modify implementation to provide a generic grouping mechanism.
A span is allocated to a group based on the `spanGroupKey` configuration
object.

Default behaviour is to use the `serviceName` as the key, but the user
can override this and either specify a preset, or a list of tags that
will be used, in order.

Use this key where span grouping and colouring decisions are made in the
UI.