Store preferred span name column width in localStorage
Different developers tend to look at different kinds of traces, meaning
that they often prefer some set width of span name column in trace view.

Without this patch, the width is set at 25% and it resets on refresh.

With this patch the width is remembered in localStorage and preserved
on refresh, making it less necessary to adjust.

Signed-off-by: Ivan Babrou <github@ivan.computer>