Update TypeScript and ESLint to latest versions (#1256)
## Which problem is this PR solving?
- Resolves #818

## Short description of the changes
Update TypeScript, ESLint and `@typescript-eslint` to the latest
versions.

Notable changes:
* Update the list of core ESLint rules that need to be disabled because
`@typescript-eslint` provides TypeScript-specific equivalents for them,
and change existing suppressions referencing the now-disabled core
rules. Group these together in the ESLint config for clarity.
* Fix references to unimported types in `serviceGraph.tsx` and
`types.tsx`.
* It seems that the type of `defaultConfig` and the user-defined config
wasn't correctly inferred as `Config` before the upgrade. Fix various
type errors resulting from this:
* `'__mergeFields'`, i.e. the list of config options where
user-specified values should be merged with the default object values in
`defaultConfig`, was defined via `defineProperty` and so seemingly
inferred as an unknown type. This now causes a type error when iterating
over its keys to try and merge the relevant property values, as
TypeScript can't guarantee that the property values are objects. Make it
a separate named export instead, which also allows for strictly typing
the allowable property names.
* Make the `url` field of the `ConfigMenuItem` type optional as this was
now recognized as a type error. This matches documentation and actual
usage.
* Rename the `dagMaxServicesLen` property definition to
`dagMaxNumServices` as that is what the actual code uses.
* Fix the `linkPatterns` config option typedef to be an array, as per
actual usage.
* Update `handleError` in Plexus to correctly take either an
`ErrorEvent` or a `MessageEvent` to match updated typings for the Web
Worker API.
* Update `catch` clauses in `readJsonFile` because TypeScript 4.4+
[types the error in `catch` clauses as `unknown` by
default](https://devblogs.microsoft.com/typescript/announcing-typescript-4-4/#using-unknown-in-catch-variables).
We don't expect anything outside of an `Error` to be thrown here, so we
can restore the previous behavior.

Signed-off-by: Máté Szabó <mszabo@fandom.com>