Remove unused Redux selectors (#1284)
## Which problem is this PR solving?
- Contributes towards making coverage reports more consistent

## Short description of the changes
Some of the inconsistency of coverage results is due to randomly
generated test data in trace.test.js. While investigating, however, I
noticed that most of the Redux selectors in this directory were only
referenced by each other and by tests - it seems they were used at some
point (they're in use in the initial commit) but were progressively
removed by refactorings. Since selectors only serve to encapsulate and
reuse logic to lookup data from the Redux store[1], they are not useful
when not imported. As such, take this opportunity to remove the dead
code.

---
[1] https://redux.js.org/usage/deriving-data-selectors

Signed-off-by: Máté Szabó <mszabo@fandom.com>