Render HTTP header tags as comma-separated strings (#1295)
## Which problem is this PR solving?

Resolves #1294

## Short description of the changes

Change `formatValue` to check whether the (de-serialized) value is an
array and, if so, whether the key should be rendered as a
comma-separated list of strings. When both conditions are true, render
the value like so:

```html
<div class="json-markup">
  <span class="json-markup-string">value1</span>, <span class="json-markup-string">value2</span>
</div>
```

![image](https://user-images.githubusercontent.com/62573/227103539-aeaf81cd-d5e3-4d4e-a878-f4f68a1503e8.png)

Currently, tag keys starting with `http.request.header.` and
`http.response.header.` render as string lists. Please let me know if
there are any other attributes I should include.

---------

Signed-off-by: Will Rogers <wjrogers@gmail.com>
Co-authored-by: Yuri Shkuro <yurishkuro@users.noreply.github.com>