Escape keys/values when rendering JSON objects (#1498)
The json-markup library does not properly escape the keys. Not clear if
it is maintained or not, last changes were several years ago. Since it's
just a single file, copy it directly and fix in place:
  * escape keys
  * parse URL and use safe href property in `<a href="...">`
* remove handling of `date` type which will never be in the JSON that we
obtain from span tags via JSON.parse()
  * remove skinning ability since we explicitly define css

Add tests:
* use Jest snapshot to validate that the JSON is escaped correctly,
which was not tested by the previous code
* use more complex JSON with different unsafe values and exercising all
code paths of the ported code

---------

Signed-off-by: Yuri Shkuro <ysh@meta.com>