Add depcheck for CI step to check for unused dependencies (#1677)
## Which problem is this PR solving?
- Fixes #1672 

## Description of the changes
- This PR adds a step in the unit-test workflow to check for unused
dependencies using the depcheck package.
- Certain dependencies which are installed but still being given in the
output are put into `.depcheckrc.json` file to be ignored.
- The `--skip-missing` flag is also added to ignore the missing deps
warning

## How was this change tested?
- The `ignores` array without any dep in `.depcheckrc.json` file was
pushed to the branch for testing, and it was observed that the CI
failed, which is the required behaviour: Workflow Run:
https://github.com/anshgoyalevil/jaeger-ui/actions/runs/5854765474/job/15871258991
- The `ignores` array with all unused deps was pushed for CI testing,
and the CI passed the Depcheck flow. Workflow Run:
https://github.com/anshgoyalevil/jaeger-ui/actions/runs/5854816286/job/15871414318#step:6:10
(Ignore the Lint fail error, as it is resolved in this PR, but workflow
run is not available because target push branch was changed to `main` in
that commit, and thus workflow run is not available to be shared)

## Notes:
Before merging this PR, all the unused deps which are neither added to
the `ignores` array, not have been removed need to be removed. Thus PR
#1675 needs to be merged before this one.
After that, this PR's branch needs to be updated with that code so that
the Depcheck doesn't fail in CI for this or for upcoming PRs.

## Checklist
- [x] I have read
https://github.com/jaegertracing/jaeger/blob/master/CONTRIBUTING_GUIDELINES.md
- [x] I have signed all commits
- [x] I have added unit tests for the new functionality
- [x] I have run lint and test steps successfully
  - for `jaeger`: `make lint test`
  - for `jaeger-ui`: `yarn lint` and `yarn test`

---------

Signed-off-by: Ansh Goyal <anshgoyal1704@gmail.com>
Signed-off-by: dependabot[bot] <support@github.com>
Signed-off-by: Pavol Loffay <p.loffay@gmail.com>
Signed-off-by: wck-iipi <21dcs006@nith.ac.in>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
Co-authored-by: Pavol Loffay <p.loffay@gmail.com>
Co-authored-by: Yuri Shkuro <yurishkuro@users.noreply.github.com>
Co-authored-by: Wck-iipi <110763795+Wck-iipi@users.noreply.github.com>