Upgrade ant-design to v5.x (#1907)
## Which problem is this PR solving?
- part of: https://github.com/jaegertracing/jaeger-ui/issues/998
- fixes https://github.com/jaegertracing/jaeger-ui/issues/1703

## Description of the changes
- The Size of this PR may look big, but there were no breaking changes
except the following:
- ant-design v5.0.0 no longer supports less variables. That means we
have to use the `seedToken` and `aliasToken` in place of less variables.
- With this change, we also needed to use the default styling of
ant-design v4. For that, ant-design team provides a `compatible`
package.
- The notification.`close` and `warn` have been renamed `destory` and
`warning` respectively.
- PR: https://github.com/jaegertracing/jaeger-ui/pull/1839/files is
reversed in this PR, because we are not really on @types/react v18. We
are locked on v16 due to yarn resolution. Having v18 written in
package.json was creating conflict with the installation of antd v5

[screen-capture
(5).webm](https://github.com/jaegertracing/jaeger-ui/assets/94157520/550a79b7-b333-497b-9b73-70d0e6a1a748)

## How was this change tested?
- Manually, and unit tests

## Checklist
- [x] I have read
https://github.com/jaegertracing/jaeger/blob/master/CONTRIBUTING_GUIDELINES.md
- [x] I have signed all commits
- [x] I have added unit tests for the new functionality
- [x] I have run lint and test steps successfully
  - for `jaeger`: `make lint test`
  - for `jaeger-ui`: `yarn lint` and `yarn test`

---------

Signed-off-by: Ansh Goyal <anshgoyal1704@gmail.com>