Fix broken Hot Reload of Plexus Package (#2089)
Addresses Hot Reload Dysfunctionality in Plexus Package: Unresponsive to
src Package Changes in UI

Implemented comprehensive modifications to rectify the malfunctioning
Hot Reload feature in the Plexus package. This issue resulted in an
inability to reflect changes made to the src package in the user
interface (UI). The following refinements have been executed to mitigate
the problem:

1. Conducted a meticulous analysis to identify and address the root
cause of the Hot Reload dysfunction within the Plexus package.
2. Incorporated strategic adjustments package.json and individual
imports as to point to the right directory instead of lib
3. Rigorously tested the implemented changes to validate their efficacy,
with a particular focus on ensuring real-time reflection of src package
modifications in the UI.

Testing Process:

The remedial changes underwent a robust testing process to ensure a
comprehensive resolution to the Hot Reload issue:

- Confirmed the restoration of expected Hot Reload functionality,
facilitating the prompt reflection of changes made to the src package in
the UI.

## Checklist

To comply with the project's contribution guidelines, the following
checklist items have been addressed:

- ✓ I have carefully read and followed the [contributing
guidelines](https://github.com/jaegertracing/jaeger/blob/master/CONTRIBUTING_GUIDELINES.md).
- ✓ I have signed all commits.
- ✓ Unit tests for the new functionality have been added.
- ✓ Linting and testing steps have been successfully executed:
  - for `jaeger`: `make lint test`
  - for `jaeger-ui`: `yarn lint` and `yarn test`

Your constructive feedback is welcome and appreciated. Thank you for
your time and consideration.

---------

Signed-off-by: Wise-Wizard <saransh.shankar@gmail.com>
Signed-off-by: Saransh Shankar <103821431+Wise-Wizard@users.noreply.github.com>
Co-authored-by: Yuri Shkuro <yurishkuro@users.noreply.github.com>