[fix]: Remove redux-form dependency from sort selector (#2569)
## Which problem is this PR solving?
- part of #2556 

## Description of the changes
- This PR removes the dependency of Sort Selector on redux-form.
- We could previously sort the traces on homepage, and the value we were
selecting on this Sort Selector was reflecting to the SearchTrace page
that is responsible to rendering the Sorted Traces list and sorting the
traces. This state exchange was previously done by redux-form but now we
are directly passing it using react state and component props.
- We were also tracking the changes done by a user in this sort selector
using a middleware tied to the redux-actions. Now we are not using redux
for form state management so tracking logic is directly moved to the
part where a sort action takes place, i.e., `handleSortChange`
- Removed @types/redux-form as there is no .ts/x file left that is using
redux-form.

## How was this change tested?
- npm run test
- npm run start and using the sort selector

## Checklist
- [x] I have read
https://github.com/jaegertracing/jaeger/blob/master/CONTRIBUTING_GUIDELINES.md
- [x] I have signed all commits
- [x] I have added unit tests for the new functionality
- [x] I have run lint and test steps successfully
  - for `jaeger`: `make lint test`
  - for `jaeger-ui`: `npm run lint` and `npm run test`

---------

Signed-off-by: Hariom Gupta <guptahariom03082003@gmail.com>
Co-authored-by: Yuri Shkuro <yurishkuro@users.noreply.github.com>