test: add test coverage for button click handler  (#2788)
This PR improves test coverage in the MonitorATMEmptyState component by
mocking the getConfigValue function to include a button with an onClick
handler. The component is then rendered with this mocked configuration,
and a simulated button click is performed using fireEvent.click. The
test asserts that the onClick handler is invoked as expected. As a
result of this addition, the test coverage for the component increased
from 75% to 100%.

Closes #2785

Signed-off-by: Vishvamsinh Vaghela <vaghelavishvamsinh11111@gmail.com>