Remove 10s timeout from test (#2939)
* Replace waitFor with act(), and remove where it's not needed
* For now add and keep a timing function in case these tests are failing
again in the future, to get more information. In local tests the render
typically takes up to 500ms, but not more.

When adding console logging I observed that the test seems to be also
pausing right after all the statements are executed. I added an explicit
cleanup() after each test to see if it helps.

Ran the test 100 times (just this one, using `it.once()`), it never
failed.
```
set -ex

rm -rf log.txt
touch log.txt
for i in {1..100}; do
  echo "Running test iteration $i..." | tee -a log.txt
  npm run test-ci src/components/TracePage/TraceStatistics/index.test.js >> log.txt 2>&1
done
```

Signed-off-by: Yuri Shkuro <github@ysh.us>